requirements.txt
Cython>=0.22
nose>=1.3.4
numpy>=1.9.2
pandas>=0.15.2
PyYAML>=3.11
scipy>=0.15.1
Shapely>=1.5.6
